Regression, Classification & Algorithmic Complexity
This Jupyter Notebook provides an in-depth tutorial on Regression, Classification, and Algorithmic Complexity with Python. The notebook covers a wide range of topics starting from the basics of Supervised Learning to advanced topics such as Model Complexity and Dataset size, Data Exploration, Multivariate Analysis, KNN for Regression, Scaling Variables, Algorithm Qualification, and Scoring Linear Algorithms. The notebook also includes exercises, quizzes, and reusable functions for training and testing algorithms. Whether you are new to Machine Learning or an experienced data scientist, this notebook is a great resource for learning and practicing Regression, Classification, and Algorithmic Complexity with Python.