Supporting Python functions
This python file contains several functions for data preparation and machine learning models that are used in various notebooks uploaded in this repository. 

The first function bunch2df takes the dataset in the bunch format and converts it into a Pandas dataframe with feature names and target names.

The second function class1_binary_percent computes the percentage of positive values from a binary array containing two different values.

The third function get_X_y takes a dataframe as input and returns two arrays: X for features and y for target. The target is supposed to be the last column.

The fourth function binary_training_and_testing takes a dataframe, percentage of train size, a classifier, target name 0, target name 1, and a verbose flag as input. It splits the dataset into train and test datasets, trains the classifier, and returns the training and testing scores. If the verbose flag is set to True, it also prints the scores.

All these functions are essential for data preparation and training a binary classification model.