Linear Models, Regression, and Classification
This notebook is a comprehensive guide that covers a wide range of topics related to data science and artificial intelligence. It is designed to help beginners and professionals alike, to learn and understand the fundamental concepts of these fields.

The notebook covers topics such as linear models for regression, classification, and multi-class classification. It discusses techniques such as ordinary least square error, ridge regression, and lasso regression, and how these techniques can be used to solve real-world problems.

The notebook also provides examples and practical exercises that help users to gain hands-on experience in solving data science and artificial intelligence problems. It includes multiple datasets such as the wave dataset, Boston extended dataset, and breast cancer dataset, which are used to illustrate various concepts and techniques.

Furthermore, the notebook includes quizzes and exercises to test users' knowledge and understanding of the topics covered in the notebook. This helps users to reinforce their learning and gain confidence in applying the concepts and techniques learned in the notebook.